backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# HTTP client for external APIs
httpx==0.25.2

# Google AI/Gemini integration
google-generativeai==0.3.2

# PDF processing
PyPDF2==3.0.1

# Environment variables
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0

# CORS middleware
fastapi-cors==0.0.6

# Security
python-multipart==0.0.6

# Logging
loguru==0.7.2

# Redis for rate limiting
redis==5.0.1

# NLP Libraries
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
langdetect==1.0.9
unidecode==1.3.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
